Reorganize parallel domains
We let only domain templates in domains.py as requires by #107.
We put utility classes for parallel computations in a dedicated module
parallel_domains.py
